backend/requirements.txt
fastapi
uvicorn[standard]
motor==3.5.1
pydantic
python-dotenv
gunicorn
pymongo==4.9.2
